+++ XMapper.Tests/TestHelpers/DoesNotThrow.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace XMapper.Tests;
+
+public static class Does
+{
+    public static void NotThrow(Action action)
+    {
+        action();
+    }
+}